backend/requirements.txt
six==1.15.0
pandas==1.0.5
numpy==1.18.5
Django==3.1.10
djangorestframework==3.12.4
